tests/krate/publish: Extract `categories` module
